+++ src/commands/replayBlock.ts
@@ -1,0 +1,5 @@
+import { ReplayBlockOptions } from "../types";
+
+export async function replayBlock(_options: ReplayBlockOptions) {
+  throw new Error("Not implemented");
+}
