--- src/modes/index.ts
@@ -1,2 +1,0 @@
-export * from "./replay";
-export * from "./run";
