--- src/modes/replay.ts
@@ -1,9 +1,0 @@
-import { ReplayBlockOptions, ReplayTxOptions } from "../types";
-
-export async function replayBlock(_options: ReplayBlockOptions) {
-  throw new Error("Not implemented");
-}
-
-export async function replayTx(_options: ReplayTxOptions) {
-  throw new Error("Not implemented");
-}
